+++ src/main.rs
@@ -6,7 +6,7 @@ use std::path::PathBuf;
 use clap::builder::ArgPredicate;
 use clap::{ArgAction, Parser, ValueEnum};
 
-use anyhow::{anyhow, Context, Result};
+use anyhow::{anyhow, bail, Context, Result};
 
 use const_format::formatcp;
 
@@ -15,6 +15,11 @@ use thiserror::Error as ThisError;
 use terminal_size::terminal_size;
 
 use hexyl::{Base, BorderStyle, CharacterTable, Endianness, Input, PrinterBuilder};
+
+use hexyl::{
+    COLOR_ASCII_OTHER, COLOR_ASCII_PRINTABLE, COLOR_ASCII_WHITESPACE, COLOR_NONASCII, COLOR_NULL,
+    COLOR_RESET,
+};
 
 #[cfg(test)]
 mod tests;
@@ -179,6 +184,10 @@ struct Opt {
         conflicts_with("panels")
     )]
     terminal_width: Option<NonZeroU64>,
+
+    /// Print a table showing how different types of bytes are colored.
+    #[arg(long)]
+    print_color_table: bool,
 }
 
 #[derive(Clone, Debug, Default, ValueEnum)]
@@ -231,10 +240,21 @@ impl From<GroupSize> for u8 {
 fn run() -> Result<()> {
     let opt = Opt::parse();
 
+    if opt.print_color_table {
+        return print_color_table().map_err(|e| anyhow!(e));
+    }
+
     let stdin = io::stdin();
 
     let mut reader = match opt.file {
-        Some(filename) => Input::File(File::open(filename)?),
+        Some(filename) => {
+            if filename.is_dir() {
+                bail!("'{}' is a directory.", filename.to_string_lossy());
+            }
+            let file = File::open(&filename)?;
+
+            Input::File(file)
+        }
         None => Input::Stdin(stdin.lock()),
     };
 
@@ -409,10 +429,9 @@ fn run() -> Result<()> {
 
     let character_table = opt.character_table;
 
-    let stdout = io::stdout();
-    let mut stdout_lock = BufWriter::new(stdout.lock());
+    let mut stdout = BufWriter::new(io::stdout().lock());
 
-    let mut printer = PrinterBuilder::new(&mut stdout_lock)
+    let mut printer = PrinterBuilder::new(&mut stdout)
         .show_color(show_color)
         .show_char_panel(show_char_panel)
         .show_position_panel(show_position_panel)
@@ -466,6 +485,42 @@ impl From<NonNegativeI64> for u64 {
         u64::try_from(x.0)
             .expect("invariant broken: NonNegativeI64 should contain a non-negative i64 value")
     }
+}
+
+fn print_color_table() -> io::Result<()> {
+    let mut stdout = BufWriter::new(io::stdout().lock());
+
+    writeln!(stdout, "hexyl color reference:\n")?;
+
+    // NULL bytes
+    stdout.write_all(COLOR_NULL)?;
+    writeln!(stdout, "⋄ NULL bytes (0x00)")?;
+    stdout.write_all(COLOR_RESET)?;
+
+    // ASCII printable
+    stdout.write_all(COLOR_ASCII_PRINTABLE)?;
+    writeln!(stdout, "a ASCII printable characters (0x20 - 0x7E)")?;
+    stdout.write_all(COLOR_RESET)?;
+
+    // ASCII whitespace
+    stdout.write_all(COLOR_ASCII_WHITESPACE)?;
+    writeln!(stdout, "_ ASCII whitespace (0x09 - 0x0D, 0x20)")?;
+    stdout.write_all(COLOR_RESET)?;
+
+    // ASCII other
+    stdout.write_all(COLOR_ASCII_OTHER)?;
+    writeln!(
+        stdout,
+        "• ASCII control characters (except NULL and whitespace)"
+    )?;
+    stdout.write_all(COLOR_RESET)?;
+
+    // Non-ASCII
+    stdout.write_all(COLOR_NONASCII)?;
+    writeln!(stdout, "× Non-ASCII bytes (0x80 - 0xFF)")?;
+    stdout.write_all(COLOR_RESET)?;
+
+    Ok(())
 }
 
 #[derive(Clone, Copy, Debug, Default, Hash, Eq, Ord, PartialEq, PartialOrd)]