+++ CHANGELOG.md
@@ -1,3 +1,21 @@
+# unreleased 
+
+## Features
+
+## Bugfixes
+
+
+# v0.16.0
+
+## Features
+
+* New `--print-color-table` option, see #229 (@sahinfalcon)
+
+## Bugfixes
+
+- Throw an error when try to view a directory, see #234 (@Integral-Tech)
+
+
 # v0.15.0
 
 ## Features